DAT-591: fix prod build — split server IO out of the client-imported graph module
CI cockpit build failed: TanStack import-protection caught the client canvas
importing operating-model-graph.ts, whose top-level pulled in db/metadata/client
(server-only bun SQL). Unit gates/typecheck don't catch this — only the build does.

Split into:
- operating-model-graph.ts — PURE (types + buildOperatingModelGraph +
  computeVisibleGraph), imports only the pure display-names; inlines the driver
  dimension projection (drops the look-drivers import). Client-safe.
- operating-model-load.ts — SERVER-only IO (loadOperatingModelGraph): metadataDb +
  config + look_* reads. Reached only via the route's createServerFn.

Route imports the loader from operating-model-load; canvas imports the pure module.
Builder test no longer needs the DB/config mocks. Verified: bun run build ✓,
1262 tests, tsc, biome all green.
